Split Location out into its own file.
This is a somewhat crude split which we'll probably refine over time, but the
mt.rs file was getting too crowded and hard to navigate. This commit contains no
functional changes.
